jwt_and_user_routes

+++ backend/controllers/goalsController.js
@@ -1,66 +1,90 @@
 import asyncHandler from "express-async-handler"
 import Goal from "../models/goalModel.js"
+import User from "../models/userModel.js"
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
-const getGoals = asyncHandler (async(req,res) =>{
-    const goals = await Goal.find()
-    res.status(200).json(goals)
+const getGoals = asyncHandler(async (req, res) => {
+  const goals = await Goal.find({ user: req.user.id })
+  res.status(200).json(goals)
 })
 
 // @desc    Set goal
 // @route   POST /api/goals
 // @access  Private
 const setGoal = asyncHandler(async (req, res) => {
-    if (!req.body.text) {
-        res.status(400)
-        throw new Error('Please add text');
-    }
-    const goal = await Goal.create({
-        text: req.body.text
-    })
- 
-    res.status(200).json(goal)
+  if (!req.body.text) {
+    res.status(400)
+    throw new Error("Please add text")
+  }
+  const goal = await Goal.create({
+    text: req.body.text,
+    user: req.user.id,
+  })
+
+  res.status(200).json(goal)
 })
-
-
 
 // @desc    Update goal
 // @route   PUT /api/goals/:id
 // @access  Private
-const updateGoal = asyncHandler(async(req,res) =>{
-    const goal = await Goal.findById(req.params.id)
+const updateGoal = asyncHandler(async (req, res) => {
+  const goal = await Goal.findById(req.params.id)
 
-    if(!goal){
+  if (!goal) {
     res.status(400)
-    throw new Error('Goal not found')
-    }
+    throw new Error("Goal not found")
+  }
+  const user = await User.findById(req.user.id)
 
-    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-      })
-      res.status(200).json(updatedGoal)
+  // Check for user
+  if (!req.user) {
+    res.status(401)
+    throw new Error("User not found")
+  }
+
+  // Make sure the logged in user matches the goal user
+
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error("User not authorized")
+  }
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+  })
+  res.status(200).json(updatedGoal)
 })
-
 
 // @desc    Delete goal
 // @route   DELETE /api/goals/:id
 // @access  Private
-const deleteGoal = asyncHandler(async(req,res) =>{
-    const goal = await Goal.findById(req.params.id)
+const deleteGoal = asyncHandler(async (req, res) => {
+  const goal = await Goal.findById(req.params.id)
 
-    if(!goal){
+  if (!goal) {
     res.status(400)
-    throw new Error('Goal not found')
-    }
+    throw new Error("Goal not found")
+  }
 
-    await Goal.deleteOne({ _id: goal._id });
+  const user = await User.findById(req.user.id)
+
+  // Check for user
+  if (!req.user) {
+    res.status(401)
+    throw new Error("User not found")
+  }
+
+  // Make sure the logged in user matches the goal user
+
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error("User not authorized")
+  }
+
+  await Goal.deleteOne({ _id: goal._id })
 
   res.status(200).json({ id: req.params.id })
 })
 
-
-export {
-    setGoal,getGoals,updateGoal,deleteGoal
-}
+export { setGoal, getGoals, updateGoal, deleteGoal }
